20180622 final

+++ src/main/java/org/zgdf/ea/controller/CustomersDeleteController.java
@@ -1,6 +1,7 @@
 package org.zgdf.ea.controller;
 
 import java.io.IOException;
+import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
@@ -14,15 +15,20 @@ public class CustomersDeleteController extends HttpServlet {
             throws ServletException, IOException {
 
         //System.out.println(request.getParameter("id"));
-        
         int id = Integer.parseInt(request.getParameter("id"));
-        
+
         CustomersDao dao = new CustomersDao();
 
-        dao.delete(id);
+        if (dao.hasActivity(id)) {
+            request.setAttribute("vanmunka", "Az ügyfélhez tartozik munka, ezért nem törölhető!");
+            RequestDispatcher rd = request.getRequestDispatcher("CustomersListController");
+            rd.forward(request, response);
+        } else {
+            dao.delete(id);
 
-        response.sendRedirect("CustomersListController");
-        
+            response.sendRedirect("CustomersListController");
+        }
+
     }
 
 }
